requirements.txt
alpaca-py>=0.26.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
pyyaml>=6.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
pytz>=2023.3
requests>=2.31.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
